requirements.txt
paho-mqtt==1.4.0
ruuvitag-sensor==0.13.0
supervisor==4.0.4